api-tests/graphql-ecommerce/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.216.1
